SmrSession: improve getRequestVar for strict types
* getRequestVar returns a string
* getRequestVarInt returns an int
* getRequestVarIntArray returns an array of ints

These are now thin wrappers around the `Request::getVar*`, which has
a few important differences than the previous implementation:

The index must exist in either $var or $_REQUEST. If it is found in
both, then an exception will be raised (previously, it would return
the value in $_REQUEST). If it is found in neither, then an exception
will also be raised (previously, it would return null).

This ensures that we are precise about our var data sources.