trader_search_result.php: make player_name optional
This page is also used as the action for clicking on a player name,
and in this pathway we only provide the `player_id`, so we need to
supply a default value (the empty string) for `player_name`.

Fixes a bug introduced in 6b0a05da68, when we made `getRequestVar`
raise an exception when the index does not exist.