account_edit.php: fix Request error
When assigning ban points to a player, if the 'suspicion' field was
disabled (i.e. for non-multi bans), then the field would not be
submitted and we would get a Request error.

We no longer use the 'suspicion' field for just the multi "Match list"
info anyway, so instead of disabling it for non-multi bans, suggest
that the admin add some private details about the ban in a placeholder.

The field now has a `maxlength=255` so that we don't accidentally
overflow the database (this field is a `varchar(255)`).