Throw user error on invalid social login type (#837)
Because the social login type should be internally consistent between
the URL on the login page and the types accepted by the SocialLogin
class, we were raising an exception if the type passed in the URL to
the login_social_processing.php page was unknown. This helped ensure
that we would be notified (by bugs emails) if we accidentally broke
social logins.

Unfortunately, because the login_social_processing.php page is linked
on the public-facing login page, it has been added to a bunch of bot
databases. Most of the search engine spiders respect our robots.txt
file, but some do not, and SQL injection attacks obviously do not.

As a result, we get a lot of unwanted bugs emails where a bot uses an
outdated or malicious social login type. This is just too much noise,
and so we now redirect to error.php if the social login type is not
valid.