Fix offline and disabled redirection (#838)
The login.inc engine page was deleted when the login skeleton was
added in 64a7db74ff3, but this was not reflected in the pages for
disabled accounts and when the game is offline, causing an error.

The forwarding mechanism for disabled.php was also broken by strict
typing, due to an SN not existing at the time it was forwarded.

Rather than refactor login.php so that it can be used in the offline
and disabled pages, we will now explicitly redirect back to the
login page. We use a session to store the disabled/offline message
(instead of putting it in the URL) since these messages are dynamic
and we don't want to have to worry about the URL length limit.

NOTE: Added the year to DEFAULT_DATE_FULL_LONG so that there is no
ambiguity in when a disabled account reopens.