Facebook.class.php: don't override default version
The `default_graph_version` defaults to the latest version available
in the loaded `facebook/graph-sdk` module. We will therefore pin the
version we want in composer.json and not override the default in our
SocialLogin::Facebook class.

As a result of this change, we will be using the v5.5 API instead of
v2.12 (which will reach EOL on 2020-05-05).