smr.inc: catch Throwable instead of Exception
In smrBBCode, we have a broad try-catch around the processing of
BBcode tags, like `[alliance=1]`. With strict-type checking, we are
often throwing more kinds of errors, so we need to broaden the catch
to include all errors (Throwable is the base class of Exception).

Ideally we would not use such a broad try-catch (because it spans
code that should *not* throw an error), and only locally handle
invalid input. (Something for the future...)