Fix error during login creation
When creating a login, we don't have an SN yet, which causes the call
to `SmrSession::update` to fail when we're using `forwardURL`. In this
case, we do not need to update the session, and since this is the only
place where `forwardURL` is used, we refactor the forwarding functions
so that `forward` updates the session and `forwardURL` does not.