Fix sector BBLinks
These links were not working due to the sectorID being saved in `$var`
as a string instead of an int, which was an issue after strict typing
was enabled.

We now convert the value to an int if it is supposed to be numeric,
and slightly modified the logic for the BBCODE_CHECK pass.

The `id` tag is no longer available. For example:

[player id=3 showalliance=true]

should now just be

[player=3 showalliance=true]