Fix error when changing CSS settings
Prior to 73d0eea529, the check that the colour scheme was valid was
artificially passing, regardless of the input to `setColourScheme`.
This was because `in_array` will do implicit type-conversion when
`strict=false` (the default), and strings will convert to `int(0)`!
Thus, the following is always true:

in_array('some string', [0,1,2]) === true

The underlying cause was that we were using `array_keys` on
`Globals::getAvailableColourSchemes`, but we actually wanted to use
the array values. Therefore, when we switched to `array_key_exists`
(which does *not* do implicit type-conversion!) in 73d0eea529, the
check was always failing:

array_key_exists('some string', [0,1,2]) === false

This exposes an inconsistency in PHP's behavior, but the bug in our
code is nevertheless fixed by just using `in_array` (and we add the
`strict=True` parameter to avoid this issue in the future).