planet_main.php: fix structure tooltips
The use of the `<label>` element here was intended to extend the
`<img>` tooltip to the structure name as well. However, this does
not actually work (and we were missing the `</label>` as well).

Put the title on the enclosing `<td>` element instead, which will
give the entire table cell the tooltip.