planet_ownership.php: handle multiple players claiming
If two players claim an unclaimed planet at about the same time,
then the one who clicks second will get an unexpected game error.

The reason is because the form they see will be for an unclaimed
planet, but the processing page will be for a claimed planet (since
the first player to click will process first). When processing for
a claimed planet, we require the `password` request variable, which
will not have been submitted.

This is an extreme edge case, but it would be nice to handle the
multiplayer scenarios as elegantly as possible.

There are two possible fixes:

1. Add a hidden `password` field to the "Take Ownership" form even
   in the case where the planet is unclaimed.

2. Use the empty string as the default value for the `password` when
   we query the request variable.

For the sake of preferring explicit over implicit, we choose the
former. This is safer if we, for example, change the name of the
password field in the form and forget to change it in the processing
page (otherwise it would always use the default value).