Instruct mysqli to throw an exception on error
Use the `mysqli_report` function to enable the internal error handling
in the mysqli library. Instead of returning `false` on errors, it will
now raise a `mysqli_sql_exception` instead. As a result, manual error
checking of mysqli calls is no longer needed (not only can we delete
these checks, we're now protected anywhere we had forgotten them).

We can also delete the special handling to report database errors in
our `logException` function, because the mysqli execptions include the
complete error message automatically.

This change was motivated by a subtle failure when reporting a database
error on the login_processing.php page. Because we had not instantiated
a global MySqlDatabase `$db` yet, we skipped the part of `logException`
that records the mysqli error message.