Don't store plaintext passwords in SmrSession
When a new user creates an account, we don't need to pass their login
and password to login_processing.php because we already have an active
SmrSession (so we skip the part of login_processing.php that asks for
that information).

This improves security, because the password was being saved in `$var`
in plaintext and would be output by all logging facilities (bug e-mail
and admin message box) in the event of an error.

Since we now only access the login/password from a form POST, we can
use `Request::get` instead of `Request::getVar` to query them.