login.php: make login and password required
Use the HTML5 `required` attribute to make the login form more
user-friendly.

In login_processing.php, we retain the check that these fields are
non-empty only because this is a public-facing page, and I can
imagine a scenario where bots circumvent the HTML5 check and spam
us with bug e-mails. (Right now it would just return a page with
"Password is incorrect!", but that could change in the future.)