Hide plaintext passwords from exception logging
Previously, if an error occurs on a page where the user has entered
their password, then that password would get exposed to multiple
logging facilities (bug e-mail and admin message box) when `$_REQUEST`
is exported.

To better protect sensitive data input by the user, we now mask any
`password` field in the `$_REQUEST` variable as it is being logged.