Add ability to cancel alliance invitations
This can be done from the "Invite Player" alliance option page.
It will also delete the invitation message sent to the player.

The new `SmrInvitation` class is responsible for interfacing with the
`alliance_invites_player` table.

Note that accepting one invitation no longer cancels all other open
invitations (with the notion that it's more similar to the scenario
with alliance passwords).