MySqlDatabase: add requireRecord
This new method should be used in place of `nextRecord` when we demand
that there is one (and only one) result from the active query.

In most cases, a missing record would result in some error (though
this will provide a clearer error message). However, since indexing
`null` is only a warning, sometimes this could lead to silently wrong
behavior.