Leave game when clicking on "Admin Tools" (#858)
Some admin tools are fairly time intensive (sending newsletters,
creating games, editing galaxies, etc.), and there is no reason to
be hogging a sector lock during that time.

The easiest way to avoid sector locks is to exit the current game,
if any. We generalize `game_play_preprocessing.php` to work for the
"Admin Tools" link as well.