npc.php: fix SmrSession error
NPCs don't have a real SmrSession because they don't access pages via
an SN. As a result, `retrieveVar` fails if we try to call it as part
of the NPC script's `processContainer`.

The easiest fix is to pre-emptively obtain a lock, which skips the
entire block of code in `do_voodoo` where we would normally lock and
re-init the session.

Note that this was always an issue, but it only became blocking when
strict-typing was added.