SmrPlayer: convert to unescaped names in database
Use `getDisplayName` instead of `getPlayerName` when we're displaying
the player name in HTML so that it is escaped correctly.

There are 3 contexts that we don't want to use `getDisplayName`:

1. In heading elements, since they have their own styling.

2. In HTML element attributes, since these should not be styled.

3. As input to the database or output to text-based programs.

For 1 & 2, we manually escape these with `htmlentities`. However, for
the attributes, we don't technically need to do the full HTML escaping
(we only need to escape quotes), but since DOMDocument complains about
this, we continue to fully escape for now (even though it's not really
correct, especially for data attributes or javascript).

Note: When this goes live, the `player` table needs to be manually
edited so that all player names are unescaped.

One benefit of this change is that `getPlayerByPlayerName` will now
work correctly for names that had (previously escaped) HTML entities
in it.