Show Join Chat left panel link even if not in a game
This allows players to easily join one of the chat services before
they have joined a game. If not in a game, the HOF name is used as
the default IRC nick instead of the player name.

Also de-emphasized the IRC link by moving it below the Discord widget,
and renamed the link from "IRC Chat" to "Join Chat" in the Freon22
template.