Reduce the damage and cost of level 1 weapons by 30%
This was needed to bring the efficiency of these weapons into line
with the trend for all the other power levels. From levels 5 to 2,
the efficiency increased in a consistent way, but then there was a
big jump in efficiency for level 1 weapons.

More powerful level 1 weapons made sense when the number of weapons
equippable at each level was restricted. Without that restriction,
it's too effective to stack level 5 weapons and then fill any extra
hardpoints with level 1 weapons.

With this change, the optimum strategy is less clear, though we may
need a pass over all weapons to balance efficiency, since lower level
weapons are still more efficient. It also slightly reduces overall
damage potential, which is helpful because stacking level 5 weapons
resulted in ~15% more damage output.