zero out negative values in getMR routine
this will stop getMR from feeding negative values in to the accuracy routines which will stop the SMS accuracy from zeroing out when firing upon other ships.

MR will still provide a 40 some percent advantage to a standard ship in a fight with a SMS.  Which is substantial but hardly enough to change the outcome of such a fight.