Associate home sector with draft leaders
When a draft leader is assigned, they can be associated with a "Home
Sector". If specified, players will move to that sector when picked
(if they're in sector 1 to start, i.e. there are no racial HQ's) and
will return to that sector when podded.

We override `SmrPlayer::getHome` for Draft games to use this sector
if it is set; otherwise, it falls back to the normal behavior.