Combine SmrPlayer and AbstractSmrPlayer
Previously, the abstract base class for SmrPlayer was being used to
support the DummyPlayer class, which omitted all database-related
methods. However, we combine SmrPlayer and AbstractSmrPlayer into the
base class for the following reasons:

1. This scheme was not maintained, and the base class already defined
   methods that required database access.

2. The DummyPlayer class is not currently usable, because the only
   feature it was used for (Combat Simulator) is broken.

3. This splitting caused problems for function argument type hints
   that were using AbstractSmrPlayer and then calling functions on
   the argument that only exist for SmrPlayer.

Most importantly, to override the SmrPlayer class for non-default game
types, we need to be able to write a distinct SmrPlayer class for that
type, and we want to keep function overrides to a minimum. Therefore,
it makes the most sense to keep the default implementations in the
base class.

NOTE: Added an early return to `setSectorID` when the sector ID is
unchanged. This impacts how `lastSectorID` is set in this case, which
likely did not occur in proper game functioning.