SmrPlayer: fix Scrutinizer warning in doHOFSave
The `doHOFSave` function made a complicated effort to internally
handle updating the contents of the `hasHOFChanged` attribute.
This can be simplified by letting `saveHOF` (which is the only
function that calls this protected method) clear this attribute
once all of the individual database updates are complete.

The Scrutinizer warning is essentially a false positive, but it
did hint that the logic here could be improved.