SmrPlayer: make setSectorID more fault-agnostic
Pass sector ID's instead of SmrSector instances to the EnterSector
and LeaveSector actions. This allows the previous or current sectors
to be non-existent.

This could be useful, for example, if we want to use sector 0 for a
special pre-game purpose. Since this sector doesn't exist, creating
an SmrSector instance for it would raise an exception.