SmrPlayer: make getHome a non-static method
We will want to be able to modify the home sector for draft games
based on the alliance the player is in. The simplest way to access
this information is by making `getHome` a non-static method.

Since it was being used to determine which sector to start the player
in before the `player` row existed in the database, we need to use
a placeholder sector (in this case, sector 0) and then call `getHome`
once the SmrPlayer object is created.