alliance_pick.php: change conditions for pick pool
Build the pool of pickable players correctly in games where there are
no racial HQ's. In this case, all players start in Sector 1, and so
this must not make players invalid picks.

This condition was used previously because Sector 1 is where booted
players are placed. To disambiguate this case, we instead restrict
the pool to players who have not already been picked (based on the
`draft_history` table), which will make booted players ineligible.