leave_newbie.php: only drop newbies in started games
In normal games, dropping newbie protection before the start time
is fine because you can't move, and are therefore stuck in Federal
protection.

However, in special games where players might not be starting in Fed,
they could drop newbies and be podded before the game starts. Better
to just cover our bases and disable it before the game has started.