preferences_processing.php: fix race change while landed
When we move a player from one sector to another, we have to make
sure to set them as not landed, otherwise they'll try to display
"Planet Main" in a sector with no planet. This results in the
following error:

Error Message: TypeError: htmlentities() expects parameter 1 to be string, null given in /smr/lib/Default/SmrPlanet.class.php:995
Stack trace:
0 /smr/lib/Default/SmrPlanet.class.php(995): htmlentities()
1 /smr/engine/Default/planet.inc(22): SmrPlanet->getDisplayName()
2 /smr/engine/Default/planet_main.php(3): require('/smr/engine/Def...')

While modifying the player so that they're not landed is definitely
the correct fix, this error message probably could have been more
graceful.