bank_anon_detail_processing.php: remove code dup
Code was (nearly) duplicated for both "Deposit" and "Payment" actions.
We can pull out the sections that get the transaction ID and do the
logging, since they can be easily extended to both actions.

Note that I needed to change the POST action to "Payment" since that
is the appropriate enum type (even though "Withdrawal" is a much more
accurate description of the action).