Rearrange Freon22 template right panel
Align the right panel with the middle center panel, since it looks
lopsided when aligned with the top of the page.

This panel will only include the ship info. The top right panel will
now be reserved for message notifications. The nice thing about this
is that it will no longer "push down" the ship panel when the player
gets a message.