message_view.php: simplify scout message display
We do not need to display the sender when displaying scout messages
(not only because the pinging player is linked in the message, but
also because it doesn't make sense to receive the message "from" the
person who is pinging you).

This enables us to display the entire message on a single line,
which dramatically improves readability of scout messages as a whole.

Also don't display reactive actions (report, blacklist, reply) when
viewing the "Sent Messages".