SmrAccount: avoid redundant database updates
Like in other classes that wrap a database table, don't perform any
database updates in any individual setter function. This prevents
redundant updates when multiple setters are called from the same
page.

We do have to add in a few explicit calls to `update()` throughout
the code, but I think it is worth it for the gained consistency in
class design (since the performance gains, while nice, are
inconsequential).