missions.inc: simplify requirement checking
The mission 'Steps' array has a variety of placeholders for data that
needs to be determined dynamically (sector ID, drink name, etc.).
Some of the replacements were being done in SmrPlayer::rebuildMission,
and others in `checkMissionRequirements`. This split was unnecessarily
complicated, so now it is done entirely in `rebuildMission`.

Two additional simplifications:

1. Instead of replacing placeholders manually in each element of the
   mission step array, walk recursively over all (string-like) array
   leaves. (Note that `str_replace` always outputs a string, which
   is why we skip any non-string-like leaves.)

2. Remove aliases in `checkMissionRequirements`. If we want to check
   a requirement, we now need to pass in values that have the same
   keys. This fixes a bug with {Enter,Leave}Sector actions, which were
   broken by 3f6a8a87240.