Error gracefully if a mission destination is missing
In `setupMissionStep`, if the destination sector does not exist, print
a helpful error message for the player and then abandon the mission
(since it can't be completed) instead of raising an exception. This is
needed to prevent exception spam in the case where the destination is
intentionally missing (e.g. the lack of a reachable Salvene HQ in the
recent draft game).

Ideally, we probably don't want to bestow a mission in the first place
if we know it can't be completed, but given the current infrastructure
it would be a bit cumbersome and expensive to perform that check in
`getAvailableMissions`. Something for the future...