Use better OOP for Blackjack
Add classes `Deck` and `Hand`, rename `PlayingCard` to just `Card`,
and move all these classes to a `Blackjack` namespace.

This allows us to streamline actions in bar_gambling_processing.php,
like drawing cards from the deck, or checking for blackjack.

Note that the logic is still a little convoluted due to the fact that
this is a combined processing and display page, but I'm waiting to
clean that up when it is refactored into separate processing/template
pages.