Update bartering formulas
The main goal of this change is to allow traders to start earning
experience at lower relations, which should lessen the frustration
related to bartering or grinding relations.

As a result of this change, some features that were implemented to
lessen the severity of the relations^13 factor in earned experience
(e.g. "bargaining leniency") can be removed because everything is
now linear.

* Simplify the offered price formula by removing the bargaining
  leniency factor and switching to a linear dependence on relations.

* Flatten out the experience earned dependence on the bargain price
  by removing a `pow(..., 13)` modifier. The leniency factor logic
  is also removed.

* Remove the +1% profit bonus for 2000 relations. Profit is now
  capped at 1000 relations.

* Change the flavor text after a trade to give some indication as
  to how optimal the trade was (instead of just "excellent").

Thanks to @stupidnewbie for the suggestion.