news_read_advanced.php: remove player search limits
There is no benefit to limiting the number of players that are matched
by a search term. Either you get all the players that match (if it's
less than the limit), in which case the limit doesn't matter, or you
get a subset of the actual matches, which may not include the match
you actually want to see.

This limit really only impacted the dev build, where I have a lot of
players with names prefixed by "Admin". In a real game, most players
don't have degenerate names.