SmrPlayer: make relations naming more consistent
The "personal" relations (as displayed to the player) was often
referred to in the code as "pure" relations. We change the code
(function and variable names) to use "personal" instead.

Also change a few instances of "global" relations in code variable
names to "race" relations (since "global" may be confused with the
"total" relations). In the future this may be renamed consistently
to "political" relations, but for now we'll stick with "race" since
it results in fewer code changes.