Add infrastructure for Enhanced Weapons
Weapons can now have one or both of two possible enhancements:

1. Increased damage (multiplicative +5% bonus)

2. Increased accuracy (additive +3% bonus)

The magnitude of these bonuses will likely need to be tweaked as we
see how they play out in the game. The intention is for them to be
small enough that they don't become mandatory in battles, but big
enough so that people will want to invest some resources to acquire
them.

If either of these enhancements are present, a weapon will be shown
in green with a "+" for each enhancement it has.

Technical notes
===============

The SmrWeapon class is broken up into an SmrWeaponType, which maps
the `weapon_type` table into PHP, and SmrWeapon, which is an entity
usable in combat by ships/planets.

Weapon enhancements are tracked in the `ship_has_weapon` and in the
`planet_has_weapon` tables. They are only properties of the SmrWeapon
class, because they are not related to weapon "types".

A few interfaces have changed to more conveniently interact with the
new design.

* SmrWeaponType no longer has a `force` argument, since the weapon
  types do not change in the database.

* SmrShip::addWeapon and SmrPlanet::addMountedWeapon now take an
  SmrWeapon instance instead of a `weapon_type_id` to avoid adding
  arguments for the ehancements (which are properties of the SmrWeapon
  instance).

* SmrWeaponTypes::getAllWeaponTypes no longer takes a game type since
  the `weapon_types` database is constant across all games (whether
  or not it should be is a separate question).

* Added methods to SmrWeapon to access information about the weapon
  enhancements.