Limit destroying planetary mounted weapons (#890)
Mounted weapons can now only be destroyed when all the mount slots
are filled (i.e. if you destroy one, you must replace it before you
can destroy another one).

The original intention was only to allow a way to replace or upgrade
weapons, and the previous implementation was the best design I could
think of at the time. However, it allowed for mounted weapons to be
entirely depleted at no cost, which was not fair (or intended).

Some unsavory tactics are still available (e.g. you can replace the
most powerful weapons with weak weapons), but the impact is no longer
devastating and comes at a small cost.