Remove unused EMP Damage and Raid weapon attributes
These weapon features are no longer in use (I believe they were an
SMR 1.5 experiment), so remove them to avoid confusion and clutter.

We could retain them in case we want to re-implement these features
in the future, but the infrastructure for them is so minimal that it
would be just as much work with and without the existing code.

Update the SMR file version to 1.07 due to a corresponding change in
the [Weapons] section format.