Add bartender hooks for creating enhanced weapons
One enhanced weapon will now be sold from a single weapon shop every
6 hours (a number likely to be tweaked).

Players can get a hint about where the sale is taking place by giving
a bartender a sufficiently big tip. If the sale is in a different
galaxy, only the galaxy name will be given; otherwise, the exact
sector is given.

The bartender will generate a new enhanced weapon sale any time after
there is less than 1 hour before the end of the current sale (so that
players have enough time to get to the shop before the sale ends).

Note that this means players cannot randomly stumble across a shop
with an enhanced weapon sale without *at least one* player having
triggered it by tipping a bartender. (It's possible that we'll also
want entering a weapon shop to be a trigger. We'll have to see.)

Technical notes
===============

The new SmrEnhancedWeaponEvent class is responsible for handling the
maintenance of the `location_sells_special` table, which tracks where
the enhanced weapon sales take place.

The `getLatestEvent` method is used by the bartender to decide what
hint to give the player, whereas the `getShopEvents` function creates
an array of sale events that are used to override the default weapon
sales.

Some flavor has been added to the bartender related to this event.
He may get sassy if you don't give him a big enough tip.