Reduce port raid profits by 10%
Port raids were too lucrative, and this was a problem that couldn't
really be solved in universe generation (by, e.g., reducing the port
density around UNOs or removing UNOs near ports). It was common for
alliances to build up enough of a warchest in the first week or so
to afford 6+ raider ships (~100M) and then make 5x or more that in a
single PR operation. This accelerated the game to endgame too quickly
by making money almost irrelevant.

If there were some risk involved in PRing (e.g. if we reverted to
random targetting for port turrets), then such a large reward might
be warranted, but this idea was met with mixed reactions, with small
alliances objecting strongly. As such, the only recourse for now is
to reduce PR profits.