add additional bounty for newbie kills
added bounty of 200 times experience for killing poorly armed newbies (read trade ships).  no stats are awarded as part of this additional bounty as to not encourage notoriety seeking.  adjusted news feed to notify others of newbie kill and the bounty that comes along with it.

Update AbstractSmrPlayer.class.php