Lock raider-class and PSF ships for first 2 weeks
The raider-class ships (and PSF) will be unavailable during the first
two weeks of the game. This is a somewhat clumsy approach to solving
the problem that when an alliance can afford a sufficient number of
raiders (usually at least 6), then we get prematurely launched into
the "endgame" due to nearly limitless PR cash and the ability to PB
any size planet.

Note: This 2-week lock duration is expected to be tweaked in the
future, or made specifiable during universe generation. Even better
would be a way to organically unlock ship classes (through player
progression and world-building).