shop_ship.php: loop over hardware types
Instead of manually listing hardware types in the ship details view,
loop over the array of all hardware types to reduce HTML duplication.

Note that this changes the "+" and "-" for CIJSD to "1" and "0" to
avoid special casing. While this doesn't visually distinguish these
"boolean" hardware types from the other types, it is easier to read.

Also center-align the numbers for visual clarity.