Freon22-layout.css: remove double indent in lists
This 40px margin for list items was intended to be used only with
a CSS reset. However, we got rid of the CSS reset from the Freon22
templates in 13aa967f204b9535. The default user agent stylesheet
already has a 40px padding for lists (`padding-inline-start`), so
this margin was doubling the intended effect.