weapon_list.php: add locations where weapons are sold
The Weapon List page now displays a "Locations" column that lists
all the shops where each weapon is sold.

This was requested as a result of the Enhanced Weapons feature,
since it will be very helpful to know what weapons are sold at
each shop.

Note that we sacrifice a bit of visual conformity (equal height
rows) to support this feature.