Replace weapon in No Shields Inc 2
The Ik-Thorne Cluster Missile (an armour-only weapon) was being sold
at No Shields Inc 2 (a weapon shop designed to sell only shields
weapons).

We replace this weapon with the Ik-Thorne Accoustic Jammer, which is
a shields-only weapon that wasn't being sold in any shop previously.
It's likely that this was the result of a typo (the weapon IDs are
14 and 15), and that it was intended to be sold here all along.

Note that "Accoustic" is misspelled. However, it's possible that this
is a deep cut referring to its high Accuracy.