Redesign database infrastructure
All tables that have an `(account_id, game_id)` primary key will be
changed to `(player_id, game_id)`. Pretty much every instance of an
account ID is replaced with a player ID in all game-scoped code.

This is the first step in a major redesign to associate tables to the
most closely-related table, rather than having everything depend on
the high-level `account` table.

Some features that this change will help facilitate:
* Multiple players per account per game (to replace "multi accounts").
* NPCs not needing an associated account.
* Free-standing ships/weapons.