Fix discord bot commands in guild channels
For reasons that are not entirely clear (either due to a discord API
change or a change in DiscordPHP 5.0+), the `$message->author` passed
to the bot commands is no longer a `Discord\Parts\User\Member` instance
in guild (public) channels. Instead, it is now the same as in private
messages, `Discord\Parts\User\User`, which has a different attribute
structure.

This issue caused all guild channel bot commands to stop functioning
due to the `$message->author->user` attribute being non-existent for
`User` instances, but the fix restores proper functionality (for now).