Fix for author ambiguity in DiscordPHP
The `$author` can be either a `Discord\Parts\User\{User,Member}` for
a guild channel message. The DiscordPHP library (at v5.0.11) seems to
handle this in a non-deterministic way, because the intent is to have
a `Member` in a guild channel and a `User` in a direct message.

In the meantime, accept either class, and always use `$author->id` to
get the user ID (which I have been assured is the same in both cases).