Switch to phpDocumentor from phpDox
The phpDocumentor project is actively developed (phpDox was last
released in March 2019 with only 21 commits since) and has a larger
user base. Unfortunately, it does not have the equivalent of the
phpDox "git enhancer", which included git history in the docs.

We forego using the official phpdoc/phpdoc image because it is 921MB,
whereas php:7.4-cli-alpine is only 83MB. As a result, we do need to
download the 24MB phpDocumentor PHAR explicitly, but it adds minimal
complexity to the Dockerfile.

Also update the api-docs GitHub action. Note that we need a separate
Dockerfile (even though it is quite similar to the one in api-docs)
because GitHub docker container actions must follow a very specific
interface.