Dockerfile: access git hash more robustly
When HEAD is a branch, the file `.git/HEAD` contains the string
"ref: refs/heads/<branch name>". However, if HEAD is in a detached
state, it only contains the hash.

The previous command only worked when HEAD was a branch. The new
command works even if HEAD is in a detached state.

Note that we don't simply use `git rev-parse HEAD` because we don't
want to install git here (it is a heavyweight dependency).