Make power level 5 weapons unique
This change is made as part of an overall balance tweak to weapon
loadouts in response to the removal of power level restrictions.
It turned out that stacking P5 weapons was still too powerful, even
with the 5% reduction in accuracy.

Now, each P5 weapon is "Unique" (you can only have one of each type
of unique weapon on your ship at a time). In this implementation,
you can still have multiple P5 weapons, but not of the same type
(i.e. 1x HHG + 1x PPL allowed, but 2x HHG not allowed); but this may
need to be changed if it is considered too powerful.

To make this clearer to the players, this is displayed on the Weapon
List page as a "Restriction", even though it's not related to the
`weapon_type.buyer_restriction` database property. (Note that we had
to modify the Weapon List filtering since previously there were no
weapons with multiple restrictions.)