Modify structure limits for Protoplanets
The Protoplanet planet type was designed for "Hunter Wars" game types,
where players would not be in top racial ships. As such, it was set up
to be weaker than other planet types. However, we are now using it
consistently in standard games, and so it needs a small balance buff.

Generators: 0 -> 5
Hangars: 40 -> 50
Bunkers: 20 -> 15