Reduce strength of PT relative to HPT
This change is intended to provide a bit more incentive to make peace
with the Human race, because the neutral PT was barely worse than the
HPT (being only -5 damage).

Photon Torpedo => -1% accuracy
Human Photon Torpedo => +1% accuracy, -5 damage

The efficiency of the HPT remains the same, and the efficiency of the
PT drops a little bit (it is now the least efficient P3 neutral).

We swap 5 damage for 1% accuracy on the HPT primarily because of the
lore that Humans have focused on high accuracy weapons.