Scout ships no longer hit mines
The point of this change is to add some diversity to ship identity,
and to encourage the use of traditionally unused ships.

Note that it does impact some real gameplay tactics (e.g. parking in
a dead end is easier to scout), so we will have to monitor how this
plays out.