SmrPort: fix division by 0 edge case
If the ideal price is small enough that relations have no impact on
the offered price, then we were getting a division by 0 error when
calculating the earned exp percent.

Example: 1 unit of wood at 1x distance has both an ideal price and
an offered price (at 0 relations) of 1 credit.

To fix this, we simply ensure that the denomenator in the equation
is 1 or greater. This has no practical impact on any outcome.

This has been an issue since at least 2009 (63d602ba8cfa).