Fix port defence pumping exploit
Previously, you could buy a single unit of trade goods for an absurd
number of credits to instantaneously pump up the port defences (since
the defences are linear in the port's credits) to the point where it
is unbustable. This is considered an exploit because the cost is too
low for the power of the tactic (even though the credit cost is high,
it only takes 1 turn).

To fix this, we impose the same limit on pumping the port's credits
as we do to the port's upgrade. This will make defence pumping cost
a proportionate amount of turns and time.